switch to data collection with terminal signals

+++ arguments.py
@@ -67,7 +67,7 @@ def get_args():
                         help='port to run the server on (default: 8097)')
 
     parser.add_argument('--repeat', type=int, default=1)
-    parser.add_argument('--channel-width', type=int, default=4)
+    parser.add_argument('--channel-width', type=float, default=4)
     parser.add_argument('--collect', action='store_true', default=False)
     parser.add_argument('--collect-after', type=int, default=200,
                         help='start saving data after this many updates')
